# specs/001-phase-ii-overview/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Phase II Spec-Driven Todo Application API
  description: REST API for the Spec-Driven Todo Application with JWT authentication and user data isolation
  version: 1.0.0
  contact:
    name: Hackathon Team
    email: team@hackathon.local
servers:
  - url: https://api.todo-app.local/v1
    description: Local development server
  - url: https://api.todo-app.vercel.app/v1
    description: Production server
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with secure password hashing
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      summary: Authenticate user and return JWT token
      description: Authenticates user credentials and returns JWT access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token required for authentication. Token must be included in Authorization header as "Bearer {token}"
  schemas:
    UserRegistrationRequest:
      type: object
      required:
        - email
        - password
        - username
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          description: User's password (will be hashed)
          example: "securePassword123"
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Unique username
          example: "johndoe"
    UserLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        password:
          type: string
          format: password
          description: User's password
          example: "securePassword123"
    LoginResponse:
      type: object
      required:
        - access_token
        - token_type
        - user
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          description: Type of token
          example: "bearer"
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      required:
        - id
        - email
        - username
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        username:
          type: string
          description: User's username
          example: "johndoe"
        created_at:
          type: string
          format: date-time
          description: Timestamp when user was created
          example: "2026-01-27T10:00:00Z"
...
    Error:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid input data provided"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp when error occurred
          example: "2026-01-27T10:00:00Z"
        request_id:
          type: string
          format: uuid
          description: Request identifier for correlation (optional)
          example: "123e4567-e89b-12d3-a456-426614174000"
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Conflict - resource already exists (e.g., duplicate email)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...